requirements.txt
discord.py-self==2.0.0
selfcord==0.2.5
